backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Celery & Redis
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
